% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areal_rojstaczer_semiconfined.R
\name{areal_rojstaczer_semiconfined}
\alias{areal_rojstaczer_semiconfined}
\title{areal_rojstaczer_semiconfined}
\usage{
areal_rojstaczer_semiconfined(frequency, radius_well, transmissivity,
  storage_confining, storage_aquifer, diffusivity_confining,
  diffusivity_vadose, thickness_confining, thickness_vadose,
  loading_efficiency, attenuation, inverse = TRUE)
}
\arguments{
\item{frequency}{the frequency in cycles per time (example units: cycles per day)}

\item{radius_well}{the radius at the screened portion of the well}

\item{storage_confining}{the confining layer specific storage}

\item{storage_aquifer}{the aquifer specific storage}

\item{diffusivity_confining}{confining layer diffusivity}

\item{diffusivity_vadose}{pneumatic diffusivity of vadose zone}

\item{thickness_confining}{the confining layer thickness}

\item{thickness_vadose}{the thickness of the vadose zone}

\item{loading_efficiency}{static loading efficiency}

\item{attenuation}{the attenuation factor of the capillary fringe}

\item{inverse}{if true water level follows the inverse of a barometric pressure change}
}
\description{
areal_rojstaczer_semiconfined
}
