% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bessel_i_complex}
\alias{bessel_i_complex}
\title{bessel_i_complex}
\usage{
bessel_i_complex(x, v)
}
\arguments{
\item{x}{\code{numeric} value to evaluate}

\item{v}{\code{numeric} value to evaluate}
}
\value{
bessel function result
}
\description{
Modified Bessel function of first kind order 1
}
