% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bessel_i_complex_new}
\alias{bessel_i_complex_new}
\title{bessel_i_complex}
\usage{
bessel_i_complex_new(x, v, gamma_term)
}
\arguments{
\item{x}{\code{numeric} value to evaluate}

\item{v}{\code{numeric} value to evaluate}
}
\value{
bessel function result
}
\description{
Modified Bessel function of first kind order 1
}
