% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bessel_k}
\alias{bessel_k}
\title{bessel_k}
\usage{
bessel_k(u)
}
\arguments{
\item{u}{value for bessel function}
}
\value{
result of the bessel function
}
\description{
boost function for the Bessel function
}
