% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bessel_k_complex}
\alias{bessel_k_complex}
\title{bessel_k_complex for complex values}
\usage{
bessel_k_complex(x)
}
\arguments{
\item{x}{\code{numeric} value to evaluate}
}
\value{
bessel function result
}
\description{
Modified Bessel function of second kind order 0
}
