% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bessel_k_complex_parallel}
\alias{bessel_k_complex_parallel}
\title{bessel_k_complex_parallel}
\usage{
bessel_k_complex_parallel(x)
}
\arguments{
\item{x}{value for bessel function}
}
\value{
bessel k
}
\description{
parallel boost function for the Bessel function
}
