% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bessel_k_parallel}
\alias{bessel_k_parallel}
\title{bessel_k_parallel}
\usage{
bessel_k_parallel(x, nu)
}
\arguments{
\item{x}{value for bessel function}

\item{nu}{order}
}
\value{
bessel k
}
\description{
parallel boost function for the Bessel function
}
