% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bessel_k_single}
\alias{bessel_k_single}
\title{bessel_k_single}
\usage{
bessel_k_single(x, nu)
}
\arguments{
\item{x}{value for bessel function}

\item{nu}{order}
}
\value{
result of the bessel function
}
\description{
boost function for the Bessel function
}
