% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bh_gamma_neg}
\alias{bh_gamma_neg}
\title{bh_gamma_neg}
\usage{
bh_gamma_neg(u, a)
}
\arguments{
\item{u}{numeric to evaluate the gamma function}

\item{a}{numeric}
}
\value{
the result of the gamma function
}
\description{
Approximate the gamma function when a is negative using Boost
}
