% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bh_tgamma}
\alias{bh_tgamma}
\title{bh_tgamma}
\usage{
bh_tgamma(u, a)
}
\arguments{
\item{u}{numeric to evaluate the gamma function}

\item{a}{numeric flow dimension parameter}
}
\value{
the result of the gamma function
}
\description{
Approximate the gamma function using Boost
}
