% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bouwer}
\alias{bouwer}
\title{bouwer}
\format{data.table
\describe{
\item{datetime}{date and time}
\item{val}{depth to water in feet}
}}
\source{
\href{https://pubs.usgs.gov/of/2002/ofr02197/spreadsheets/Slug_Bouwer-Rice.xls}
}
\usage{
data(bouwer)
}
\description{
dataset for testing
}
\examples{
data(bouwer)
}
\keyword{datasets}
