% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bouwer_1976_abc}
\alias{bouwer_1976_abc}
\title{bouwer_1976_abc}
\format{data.table}
\source{
\href{https://pubs.usgs.gov/of/2002/ofr02197/spreadsheets/Slug_Bouwer-Rice.xls}
}
\usage{
data(bouwer_1976_abc)
}
\description{
a, b and c values for Bouwer and Rice, 1976 solution, for testing
}
\examples{
data(bouwer_1976_abc)
}
\keyword{datasets}
