% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bouwer_1976_water_level}
\alias{bouwer_1976_water_level}
\title{bouwer_1976_water_level}
\format{data.table
\describe{
\item{datetime}{date and time}
\item{val}{depth to water in feet}
}}
\source{
\href{https://pubs.usgs.gov/of/2002/ofr02197/spreadsheets/Slug_Bouwer-Rice.xls}
}
\usage{
data(bouwer_1976_water_level)
}
\description{
dataset for testing
}
\examples{
data(bouwer_1976_water_level)
}
\keyword{datasets}
