% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bouwer_rice}
\alias{bouwer_rice}
\title{Calculate transmissivity with Bouwer-Rice solution}
\usage{
bouwer_rice(time, drawdown, radius_screen, radius_casing, Le, Lw, H)
}
\arguments{
\item{time}{the elapsed time}

\item{drawdown}{the drawdown}

\item{radius_screen}{radius of the screen}

\item{radius_casing}{radius of the casing where the water level is}

\item{H}{}
}
\value{
transmissivity from bouwer_rice
}
\description{
Calculate transmissivity with Bouwer-Rice solution
}
