% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bouwer_rice_abc}
\alias{bouwer_rice_abc}
\title{Calculate equations 4 and 5 from bouwer, 1989}
\usage{
bouwer_rice_abc(rw, Le, Lw, H)
}
\arguments{
\item{H}{}
}
\value{
ln(Re/rw)
}
\description{
Calculate equations 4 and 5 from bouwer, 1989
}
