% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/den_iseger.R
\name{den_iseger}
\alias{den_iseger}
\title{den_iseger}
\usage{
den_iseger(impulse, time, n_weights = 16, ...)
}
\arguments{
\item{time}{must be regularly spaced}

\item{n_weights}{}
}
\description{
den_iseger
}
