% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exp_int_single}
\alias{exp_int_single}
\title{exp_int_single}
\usage{
exp_int_single(u, a)
}
\arguments{
\item{u}{value of the Theis u}

\item{a}{not used (flow dimension)}
}
\value{
exponential integral
}
\description{
Calculate the exponential integral using Boost
}
