% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution_fft.R
\name{fftw_convolve}
\alias{fftw_convolve}
\title{fftw_convolve}
\usage{
fftw_convolve(x, y, normalize = TRUE)
}
\arguments{
\item{x}{the data series}

\item{y}{the filter}

\item{normalize}{do you want the filter normalized}
}
\value{
grf solution for multiple pumping periods
}
\description{
Designed to be used for fast moving averages
}
