% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gamma_der}
\alias{gamma_der}
\title{gamma_der}
\usage{
gamma_der(u, a)
}
\arguments{
\item{u}{numeric to evaluate the gamma function}

\item{a}{numeric}
}
\value{
the result of the gamma function
}
\description{
Derivative of the gamma function.  Need to check on this.
}
