% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grf_coefficient}
\alias{grf_coefficient}
\title{grf_coefficient}
\usage{
grf_coefficient(flow_rate, radius, K, thickness, flow_dimension)
}
\arguments{
\item{flow_rate}{(vector) the flow rate}

\item{radius}{(double) distance to center of source}

\item{K}{(double) hydraulic conductivity of the fracture system}

\item{thickness}{(double) thickness}

\item{flow_dimension}{(double) flow dimension}
}
\value{
coefficient
}
\description{
Coefficient for the grf without pumping
}
