% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution_fft.R
\name{grf_convolve}
\alias{grf_convolve}
\title{grf_convolve}
\usage{
grf_convolve(radius, storativity, K, thickness, time, flow_rate,
  flow_dimension = 2, plan = NA)
}
\arguments{
\item{radius}{distance to monitoring location}

\item{storativity}{the storativity of the aquifer}

\item{K}{the hydraulic conductivity of the aquifer}

\item{thickness}{the thickness}

\item{time}{the time to output}

\item{flow_rate}{vector of flow rates}

\item{flow_dimension}{the barker flow dimension}

\item{plan}{the FFTW plan, can be missing}
}
\value{
grf solution for multiple pumping periods
}
\description{
Implementation of Barker, 1988 Generalized Radial Flow model for line source/sink.  For each time given a flow_rate needs to be provided.
}
