% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grf_parallel}
\alias{grf_parallel}
\title{grf_parallel}
\usage{
grf_parallel(a, u)
}
\arguments{
\item{a}{flow dimension}

\item{u}{value of the Theis u}
}
\value{
exponential integral
}
\description{
Parallel vector version of the exponential integral
}
