% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grf_u_time}
\alias{grf_u_time}
\title{grf_u_time}
\usage{
grf_u_time(radius, storativity, K, time)
}
\arguments{
\item{radius}{distance to monitoring interval}

\item{storativity}{aquifer storativity}

\item{K}{aquifer hydraulic conductivity}

\item{time}{prediction times}
}
\value{
u for well function
}
\description{
Calculation of grf u
}
