% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hantush}
\alias{hantush}
\title{grf}
\usage{
hantush(radius, storativity, transmissivity, leakage, time, flow_rate,
  flow_rate_times, n_terms)
}
\arguments{
\item{radius}{distance to monitoring interval}

\item{storativity}{aquifer storativity}

\item{transmissivity}{aquifer transmissivity}

\item{leakage}{hantush leakage}

\item{time}{prediction times}

\item{flow_rate}{well flow rates}

\item{n_terms}{number of terms to use in Hantush solution.  More is more precise but slower.}

\item{flow_time_interval}{time between flow rate measurements in samples}
}
\value{
theis solution for multiple pumping scenario
}
\description{
Parallel convolution of GRF well function and flow rates in the time domain.
Time series needs to be regularily spaced and so are the flow rates.  Some
performance gains can be achieved if the number of flow rate does not change
for each time.
}
