% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hantush_epsilon}
\alias{hantush_epsilon}
\title{hantush_epsilon}
\usage{
hantush_epsilon(radius, leakage)
}
\arguments{
\item{radius}{distance to monitoring well}

\item{leakage}{aquifer transmissivity}
}
\value{
coefficient Hantush well function
}
\description{
Calculation of r^2/(4B^2)
}
