% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hantush_time_parallel}
\alias{hantush_time_parallel}
\title{hantush_time_parallel}
\usage{
hantush_time_parallel(radius, storativity, transmissivity, leakage, time,
  flow_rate, flow_time_interval, n_terms)
}
\arguments{
\item{radius}{distance to monitoring interval}

\item{storativity}{aquifer storativity}

\item{transmissivity}{aquifer transmissivity}

\item{leakage}{hantush leakage}

\item{time}{prediction times}

\item{flow_rate}{well flow rates}

\item{flow_time_interval}{time between flow rate measurements in samples}

\item{n_terms}{number of terms to use in Hantush solution.  More is more precise but slower.}
}
\value{
hantush jacob solution for multiple pumping scenario
}
\description{
Parallel convolution of hantush well function and flow rates in the time domain.
Time series needs to be regularily spaced.
}
