% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hantush_well_parallel}
\alias{hantush_well_parallel}
\title{hantush_well_single}
\usage{
hantush_well_parallel(u, b, n_terms)
}
\arguments{
\item{u}{(vector) value of the Theis u}

\item{b}{the leakance}

\item{n_terms}{the number of terms used in the hantush approximation}
}
\value{
hantush well function
}
\description{
Parallel version of the hantush well function
}
