% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hsieh_1987_fig_2_3}
\alias{hsieh_1987_fig_2_3}
\title{Hsieh (1987) Figures 2 and 3 Digitized}
\format{A \code{data.table} The columns are:
\describe{
\item{\code{dimensionless_frequency}}{dimensionless frequency}
\item{\code{response}}{gain and phase of response}
\item{\code{S}}{storativity}
\item{\code{variable}}{gain or phase}
}}
\usage{
hsieh_1987_fig_2_3
}
\description{
Amplitude and phase response as a function of S.
}
\examples{
utils::data(hsieh_1987_fig_2_3)
}
\keyword{datasets}
