% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{impulse_function}
\alias{impulse_function}
\title{impulse_function}
\usage{
impulse_function(u, flow_time_interval)
}
\arguments{
\item{u}{well function}

\item{flow_time_interval}{time between flow rate measurements in samples}
}
\value{
impulse function for convolution
}
\description{
Calculation of the impulse function from a well function.
}
