% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/den_iseger.R
\name{iseger_data}
\alias{iseger_data}
\title{iseger_data}
\usage{
iseger_data(n = 16)
}
\arguments{
\item{n}{number of terms divided by 2}
}
\value{
iseger weights
}
\description{
iseger_data
}
\examples{
iseger_data(16)
}
