% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ogata_banks}
\alias{ogata_banks}
\title{Ogata-Banks solution for 1-D flow (vectorized).}
\usage{
ogata_banks(D, R, decay, v, C0, x, t)
}
\arguments{
\item{D}{doublediffusion coefficient}

\item{R}{double retardation coefficient}

\item{decay}{double decay coefficient}

\item{v}{double velocity}

\item{C0}{double concentration}

\item{x}{vector x position}

\item{t}{vector time}
}
\value{
ogata banks solution each row is an x value and each column is a time
}
\description{
Ogata, A., Banks, R.B., 1961. A solution of the differential equation of
longitudinal dispersion in porous media. U. S. Geol. Surv. Prof. Pap. 411-A.
1-D, infinite source, uniform flow, constant parameters, decay, retardation

To have values match the excel sheet
https://www.civil.uwaterloo.ca/jrcraig/pdf/OgataBanks.xlsm the decay
coefficient needs to be scaled by the retardation coefficient.
}
