% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ogata_banks_ind}
\alias{ogata_banks_ind}
\title{Ogata-Banks solution for 1-D flow.}
\usage{
ogata_banks_ind(D, v, C0, x, t)
}
\arguments{
\item{D}{diffusion coefficient}

\item{v}{double velocity}

\item{C0}{double concentration}

\item{x}{double x position}

\item{t}{double time}
}
\value{
ogata banks solution
}
\description{
Ogata, A., Banks, R.B., 1961. A solution of the differential equation of
longitudinal dispersion in porous media. U. S. Geol. Surv. Prof. Pap. 411-A.
1-D, infinite source, uniform flow, constant parameters, no decay, no retardation
}
