% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{parameters}
\usage{
parameters(frequency, period, omega, alpha_w, storage_aquifer,
  storage_confining, specific_yield, transmissivity_aquifer,
  diffusivity_vadose, diffusivity_aquifer, diffusivity_confining,
  thickness_vadose, thickness_aquifer, thickness_confining, height_water,
  radius_well, radius_casing, loading_efficiency, attenuation, inverse,
  gravity)
}
\arguments{
\item{frequency}{the frequency in cycles per time (example units: cycles per day)}

\item{period}{the period of signal (example units: days)}

\item{omega}{the angular frequency \eqn{2 \pi \omega} (example units: radians / time)}

\item{alpha_w}{the dimensionless frequency}

\item{storage_aquifer}{the aquifer specific storage}

\item{storage_confining}{the confining layer specific storage}

\item{specific_yield}{the specific yield at the water table}

\item{transmissivity_aquifer}{the aquifer transmissivity}

\item{diffusivity_vadose}{pneumatic diffusivity of vadose zone}

\item{diffusivity_aquifer}{aquifer diffusivity}

\item{diffusivity_confining}{confining layer diffusivity}

\item{thickness_vadose}{the thickness of the vadose zone}

\item{thickness_aquifer}{the aquifer thickness}

\item{thickness_confining}{the confining layer thickness}

\item{height_water}{the depth from the water table}

\item{radius_well}{the radius at the screened portion of the well}

\item{radius_casing}{the radius at the location of the water level}

\item{loading_efficiency}{static loading efficiency}

\item{attenuation}{the attenuation factor of the capillary fringe}

\item{inverse}{if true water level follows the inverse of a barometric pressure change}

\item{gravity}{the acceleration due to gravity}
}
\description{
parameters
}
