% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rojstaczer}
\alias{rojstaczer}
\title{rojstaczer}
\usage{
rojstaczer(frequency, radius_well, transmissivity, storage_confining,
  storage_aquifer, diffusivity_confining, diffusivity_vadose,
  thickness_confining, thickness_vadose, loading_efficiency, attenuation,
  gamma_term_a, gamma_term_b)
}
\arguments{
\item{attenuations}{}
}
\value{
bessel function result
}
\description{
Modified Bessel function of second kind order 0
}
