% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rojstaczer_1990_fig_4}
\alias{rojstaczer_1990_fig_4}
\title{Rojstaczer and Riley (1990) Figure 4 Digitized}
\format{A \code{data.table} The columns are:
\describe{
\item{\code{Qu}}{dimensionless frequency}
\item{\code{response}}{gain and phase of response}
\item{\code{R_div_Qu}}{ratio comparing water table drainage and unsaturated zone influences}
\item{\code{variable}}{gain or phase}
}}
\usage{
rojstaczer_1990_fig_4
}
\description{
Barometric efficiency and phase of phreatic well to atmospheric
loading as a function of dimensionless frequency Qu and R/Qu and fully
penetrating. Static barometic efficiency is 0.5.
}
\examples{
utils::data(rojstaczer_1990_fig_4)
}
\keyword{datasets}
