% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rojstaczer_parallel}
\alias{rojstaczer_parallel}
\title{rojstaczer_parallel}
\usage{
rojstaczer_parallel(frequency, radius_well, transmissivity,
  storage_confining, storage_aquifer, diffusivity_confining,
  diffusivity_vadose, thickness_confining, thickness_vadose,
  loading_efficiency, attenuation)
}
\arguments{
\item{gamma_term_b}{}
}
\value{
bessel function result
}
\description{
Modified Bessel function of second kind order 0
}
