% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rojstaczer_1988.R
\name{rojstaczer_r}
\alias{rojstaczer_r}
\title{rojstaczer_r}
\usage{
rojstaczer_r(frequency, radius_well, transmissivity, storage_confining,
  storage_aquifer, diffusivity_confining, diffusivity_vadose,
  thickness_confining, thickness_vadose, loading_efficiency, attenuation,
  inverse = TRUE)
}
\arguments{
\item{inverse}{calculate the barometric or loading response function}
}
\description{
rojstaczer_r
}
