% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slug_cbp.R
\name{slug_cbp}
\alias{slug_cbp}
\title{slug_cbp}
\usage{
slug_cbp(time, radius_casing = NULL, radius_screen = NULL,
  radius = NULL, storativity = 1e-05, transmissivity = 10,
  head_0 = 1, n = 12L)
}
\arguments{
\item{time}{elapsed time since slug introduction}

\item{radius_casing}{radius of the casing where the water level is}

\item{radius_screen}{radius of the screened interval}

\item{radius}{radius from borehole center to calculate the response}

\item{storativity}{aquifer storativity}

\item{transmissivity}{aquifer transmissivity}

\item{head_0}{height of introduced head change}

\item{n}{coefficient for Stehfest algorithm (should be < 20)}
}
\value{
vector of H/H0
}
\description{
This function calculates H/H0 for a well of finite diameter using equation 7
from Cooper, Bredehoeft and Papdopulos, 1967
Response of a Finite-Diameter Well to an Instantaneous Charge Water.

This function uses Stehfest method for the Inverse Laplace Transform.
}
\examples{
t <- 5.776 * 10^(seq(-4, 3, 0.01))
s <- slug_cbp(t, radius_screen = 7.6)
plot(s~t, log = 'x', type = 'l')
}
