% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{slug_cbp_parallel}
\alias{slug_cbp_parallel}
\title{slug_cbp_parallel}
\usage{
slug_cbp_parallel(p, radius_casing, radius_screen, radius,
  storativity = 1e-05, transmissivity = 0.01, head_0 = 1)
}
\arguments{
\item{p}{elapsed time since slug introduction}

\item{radius_casing}{radius of the casing where the water level is}

\item{radius_screen}{radius of the screened interval}

\item{radius}{radius from borehole center to calculate the response}

\item{storativity}{aquifer storativity}

\item{transmissivity}{aquifer transmissivity}

\item{head_0}{height of introduced head change}
}
\value{
cbp solution
}
\description{
parallel boost function for the Bessel function
}
