% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stehfest.R
\name{stehfest}
\alias{stehfest}
\title{stehfest}
\usage{
stehfest(time, n = 12, impulse, ...)
}
\arguments{
\item{time}{elapsed time since slug introduction}

\item{n}{coefficient for Stehfest algorithm (should be < 20)}

\item{impulse}{function to invert}

\item{...}{arguments to pass to impulse}
}
\value{
inverse laplace transform of impulse function
}
\description{
Inverse Laplace transform using Stehfest algorithm
}
