% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution_fft.R
\name{theis_convolve}
\alias{theis_convolve}
\title{theis_convolve}
\usage{
theis_convolve(radius, storativity, transmissivity, time, flow_rate,
  flow_dimension = 2)
}
\arguments{
\item{radius}{distance to monitoring location}

\item{storativity}{the storativity of the aquifer}

\item{transmissivity}{the transmissivity of the aquifer}

\item{time}{the time to output}

\item{flow_rate}{vector of flow rates}

\item{flow_dimension}{the barker flow dimension}
}
\value{
grf solution for multiple pumping periods
}
\description{
Implementation of Barker, 1988 Generalized Radial Flow model for line source/sink
}
