% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{theis_u_time}
\alias{theis_u_time}
\title{theis_u_time}
\usage{
theis_u_time(radius, storativity, transmissivity, time)
}
\arguments{
\item{radius}{distance to monitoring interval}

\item{storativity}{aquifer storativity}

\item{transmissivity}{aquifer transmissivity}

\item{time}{prediction times}
}
\value{
u for well function
}
\description{
Calculation of Theis u
}
