% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidal_hsieh_1987.R
\name{tidal_hsieh_1987}
\alias{tidal_hsieh_1987}
\title{tidal_hsieh_1987
Solution for estimating transmissivity and storativity from earth tides.}
\usage{
tidal_hsieh_1987(frequency, storativity, transmissivity, radius_well,
  radius_casing = radius_well)
}
\arguments{
\item{radius_well}{}
}
\description{
tidal_hsieh_1987
Solution for estimating transmissivity and storativity from earth tides.
}
\examples{
data('hsieh_1987_fig_2_3')
storativity <- 1e-07
transmissivity <- 1e-03
radius_well <- 0.05
frequency <- 10^seq(-5, 2, by = 0.05)
tau   <- 1 / frequency
hsieh <- tidal_hsieh_1987(frequency, storativity, transmissivity, radius_well)
plot(Mod(response)~dimensionless_frequency, hsieh,
 type='l',
 log = 'x',
 xlim = c(1, 1000))
points(response~dimesionless_frequency, hsieh_1987_fig_2_3[variable=='gain' & S == storativity])

plot(unwrap(Arg(response)) * 180/pi~dimensionless_frequency, hsieh,
 type='l',
 log = 'x',
 xlim = c(1, 1000),
 ylim = c(0, -90))
points(response~dimesionless_frequency, hsieh_1987_fig_2_3[variable=='phase' & S == storativity])

}
