% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidal_liu_1989.R
\name{tidal_liu_1989}
\alias{tidal_liu_1989}
\title{liu_1989}
\usage{
tidal_liu_1989(frequency, storativity, transmissivity, thickness_aquifer,
  height_water, radius_well, radius_casing = radius_well,
  gravity = 9.80665)
}
\arguments{
\item{gravity}{}
}
\description{
liu_1989
}
\examples{
data('liu_1989_fig_8')
storativity <- 5e-4
transmissivity <- 0.5
radius_well <- 0.117
tau   <- seq(10, 50, 0.001)
frequency <- 1/tau
liu    <- tidal_liu_1989(frequency, storativity, transmissivity, thickness_aquifer = 565, height_water = 92, radius_well, radius_casing = radius_well/2)
cooper <- tidal_cooper_1965(frequency, storativity, transmissivity, thickness_aquifer = 565, height_water = 92, radius_well, radius_casing = radius_well/2)
plot(response~period, liu_1989_fig_8, type='p')
points(Mod(response)~(tau), liu,
 type='l',
 ylim = c(0.0, 8), col = 'red')
points(Mod(response)~tau, cooper,
 type='l',
 col = 'blue')

plot(unwrap(Arg(response)) * 180/pi~tau, liu,
 type='l',
 log = 'x')
points(unwrap(Arg(response)) * 180/pi~tau, cooper,
 type='l',
 col = 'blue')
}
