% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidal_wang_2018.R
\name{tidal_wang_2018}
\alias{tidal_wang_2018}
\title{tidal_wang_2018}
\usage{
tidal_wang_2018(frequency, storativity, transmissivity, k_vertical,
  thickness_confining, radius_well, radius_casing = radius_well)
}
\arguments{
\item{radius_casing}{}
}
\description{
tidal_wang_2018
}
