% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{well_function_coefficient}
\alias{well_function_coefficient}
\title{well_function_coefficient}
\usage{
well_function_coefficient(flow_rate, transmissivity)
}
\arguments{
\item{flow_rate}{well flow rates}

\item{transmissivity}{aquifer transmissivity}
}
\value{
coefficient for Theis and Hantush well function
}
\description{
Calculation Coefficient Q/(4 pi * T)
}
