% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{well_function_convolve}
\alias{well_function_convolve}
\title{well_function_convolve}
\usage{
well_function_convolve(flow_time_interval, coefs, u)
}
\arguments{
\item{flow_time_interval}{time between flow rate measurements in samples}

\item{coefs}{well coeffiicents}

\item{u}{well impuse function}
}
\value{
impulse function for convolution
}
\description{
Parallel convolution in the time domain
}
