% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution_fft.R
\name{fftw_convolve}
\alias{fftw_convolve}
\title{fftw_convolve}
\usage{
fftw_convolve(x, y, normalize = TRUE, align = "center")
}
\arguments{
\item{x}{the data series}

\item{y}{the filter}

\item{normalize}{do you want the filter normalized}

\item{align}{what alignment}
}
\value{
convolution of data series and filter
}
\description{
Designed to be used for fast centered moving averages
}
\examples{
set.seed(37)
len <- 999
x <- cumsum(rnorm(len))
y <- rep(1, 3)
sub <- c(1, 50)
plot(1:len, x, type='o', pch = 20, xlim = sub, cex = 0.4)
points(1:len, fftw_convolve(x, y), type = 'o', col='red', pch = 20, cex = 0.4)
}
