% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidal_hsieh_1987.R
\name{fit_tidal_hsieh_1987}
\alias{fit_tidal_hsieh_1987}
\title{fit_tidal_hsieh_1987}
\usage{
fit_tidal_hsieh_1987(frequency = 1.9322736/86400, gain, phase,
  radius_well, radius_casing = radius_well)
}
\arguments{
\item{frequency}{frequency in cycles per second}

\item{gain}{value to fit in m/ns}

\item{phase}{phase shift relative to Earth tide}

\item{radius_well}{radius of the well}

\item{radius_casing}{radius of the casing}
}
\description{
This function fits a empirical response function value to the Hsieh model.
For high transmissivity wells (>1e-4 m2/sec) this will likely not give
reasonable storativity values.
}
\examples{
data('hsieh_1987_fig_2_3')
storativity <- 1e-07
transmissivity <- 1e-05
radius_well <- 0.05
frequency <- 1.9322736/86400
hsieh <- tidal_hsieh_1987(frequency, storativity, transmissivity, radius_well)

res <- fit_tidal_hsieh_1987(frequency = 1.9322736/86400,
                 Mod(hsieh$response),      # m / strain
                 Arg(hsieh$response),      # degrees
                 radius_well,
                 radius_casing = radius_well)
all.equal(res$par, log10(c(storativity, transmissivity)), tolerance = 0.05)
}
