% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grf}
\alias{grf}
\title{grf}
\usage{
grf(a, u)
}
\arguments{
\item{a}{flow dimension}

\item{u}{value of the Theis u}
}
\value{
exponential integral
}
\description{
Non parallel vector version of the exponential integral using Boost
}
