% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grf_parallel}
\alias{grf_parallel}
\title{grf_parallel}
\usage{
grf_parallel(u, a)
}
\arguments{
\item{u}{value of the Theis u}

\item{a}{flow dimension}
}
\value{
exponential integral
}
\description{
Parallel vector version of the exponential integral
}
