% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grf_time_parallel}
\alias{grf_time_parallel}
\title{grf_time_parallel}
\usage{
grf_time_parallel(radius, storativity, transmissivity, time, flow_rate,
  flow_time_interval, flow_dimension)
}
\arguments{
\item{radius}{distance to monitoring interval}

\item{storativity}{aquifer storativity}

\item{transmissivity}{aquifer transmissivity}

\item{time}{prediction times}

\item{flow_rate}{well flow rates}

\item{flow_time_interval}{time between flow rate measurements in samples}

\item{flow_dimension}{time between flow rate measurements in samples}
}
\value{
theis solution for multiple pumping scenario
}
\description{
Parallel convolution of Theis well function and flow rates in the time domain.
Time series needs to be regularily spaced.
}
