% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kdr}
\alias{kdr}
\title{kdr}
\format{data.frame}
\source{
\href{http://www.hydrology.nl/images/docs/dutch/key/Kruseman_and_De_Ridder_2000.pdf}{Kruseman_and_de_Ridder}
}
\usage{
data(kdr)
}
\description{
Data from Analysis and Evaluation of Pumping Test Data 'Oude Korendijk' (after Wit 1963)
}
\examples{
data(kdr)
}
\references{
G.P. Kruseman, N.A. de Ridder, Analysis and Evaluation of Pumping Test Data (1994), International Institute for Land Reclamation and Improvement.
}
\keyword{datasets}
