% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooper_1965.R
\name{tidal_cooper_1965}
\alias{tidal_cooper_1965}
\title{cooper_1965}
\usage{
tidal_cooper_1965(frequency, storativity, transmissivity,
  thickness_aquifer, height_water, radius_well, gravity = 9.80665)
}
\arguments{
\item{gravity}{}
}
\description{
cooper_1965
}
