% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsieh_1987.R
\name{tidal_hsieh_1987}
\alias{tidal_hsieh_1987}
\title{hsieh_1987
Solution for estimating transmissivity and storativity from earth tides.}
\usage{
tidal_hsieh_1987(frequency, storativity, transmissivity, radius_well)
}
\arguments{
\item{radius_well}{}
}
\description{
hsieh_1987
Solution for estimating transmissivity and storativity from earth tides.
}
\examples{
storativity <- 1e-04
transmissivity <- 1e-0
radius_well <- 0.05
frequency <- 10^seq(-5, 2, by = 0.05)
tau   <- 1 / frequency
hsieh_1987(frequency, storativity, transmissivity, radius_well)
}
