% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_1989.R
\name{tidal_liu_1989}
\alias{tidal_liu_1989}
\title{liu_1989}
\usage{
tidal_liu_1989(frequency, storativity, transmissivity, thickness_aquifer,
  height_water, radius_well, gravity = 9.80665)
}
\arguments{
\item{gravity}{}
}
\description{
liu_1989
}
